test: регресс-тесты не роняют сборку на слабых моделях (Assume)
Ответ на ревью linarkou (PR #129): регресс-IT не должны падать из-за качества LLM.

- Инвариант ФРЕЙМВОРКА -> жёсткий assert; кооперация/вывод модели -> Assumptions.assumeTrue
  -> тест SKIPPED (сигнал), а не fail сборки.
- Объединены GigaChatToolChainIT + GigaChatReactAgentIT -> GigaChatAgentIT
  (все сценарии сохранены: ReAct, 2-шаг, 3-шаг, селективный выбор тулов, ChatMemory);
  удалены hard-assert порядка вызовов и мёртвые поля seq/*Order.
- RAG: добавлен жёсткий инвариант ретривала без участия чата
  (similaritySearch topK=1 -> top-1 содержит код); дословный ответ модели -> assumeTrue.
- ChatClient: .entity() обёрнут (невалидный JSON слабой модели -> SKIP, не fail);
  структурный парс и стрим остаются жёсткими.

Базовая проводка тулов по-прежнему жёстко покрыта в GigaChatAutoConfigurationIT и
GigaChatModelFullIT. Прогон на GigaChat-2-Max: 10/10 регрессных IT зелёные, 0 пропусков.
